왼쪽 아래 데이터 불러오는 중

+++ src/main/java/com/korit/younggyu/plant/entity/TotalCpgEntity.java
@@ -12,5 +12,5 @@ import lombok.NoArgsConstructor;
 public class TotalCpgEntity {
     private int totalTpg;
     private int totalMpg;
-    private int totalCpg;
+    private long totalCpg;
 }
